added a different main
this main would be easier to work with when dealing with rendering a large 3d landscape than the current main, but there are still errors and things I have to work through to get it working. If I don't manage to fix the problems by the presenting date, we will just present on the current main and I will fix this one at a later date.